Move play/[scenarioId] page to backstage
